profilepicture in toolbar

+++ app/src/main/java/com/example/messenger_app_android/adapters/UserAdapter.kt
@@ -47,14 +47,14 @@ class UserAdapter(
     override fun onBindViewHolder(holder: ProfileViewHolder, position: Int) {
         val auth = FirebaseAuth.getInstance()
         val user = users[position]
-        val radius = 50
+        val radius = 50f
         holder.itemView.apply {
             display_name.text = user.displayName
             Picasso.get()
                 .load(user.profilePicture)
                 .memoryPolicy(MemoryPolicy.NO_CACHE)
                 .transform(RoundedTransformationBuilder()
-                    .cornerRadius(radius.toFloat())
+                    .cornerRadius(radius)
                     .oval(false)
                     .build())
                 .into(profile_picture)
